Increase upload max filesize limit

+++ src/main/java/com/sonicle/webtop/core/bol/js/JsUserOptions.java
@@ -86,7 +86,7 @@ public class JsUserOptions extends JsUserOptionsBase {
 	public Boolean otpDeviceIsTrusted;
 	public String otpDeviceTrustedOn;
 	
-	public Integer imUploadMaxFileSize;
+	public Long imUploadMaxFileSize;
 	public Boolean imSoundOnFriendConnect;
 	public Boolean imSoundOnFriendDisconnect;
 	public Boolean imSoundOnMessageReceived;
